Added container for creating transactions, wrote logic for rendering transaction, deleting, editing and logic for counting total sum of income, expense, added fetching transactions thunk that give transactions with sorted from last transaction, to oldest

+++ src/components/CategoryItem/CategoryItem.tsx
@@ -16,9 +16,11 @@ const CategoryItem: React.FC<Props> = ({category}) => {
 
   const onDelete = async (contactId: string) => {
     try {
-      setIsDeleting(true);
-      await dispatch(deleteCategory(contactId));
-      await dispatch(fetchCategories());
+      if (confirm()) {
+        setIsDeleting(true);
+        await dispatch(deleteCategory(contactId));
+        await dispatch(fetchCategories());
+      }
     } finally {
       toast.success('Category deleted!');
       setIsDeleting(false);
